requirements.txt
streamlit==1.29.0
audio-recorder-streamlit
python-dotenv
google-generativeai
vertexai